+++ Sora/PackageInfo.swift
@@ -9,19 +9,19 @@ public enum SDKInfo {
  */
 public enum WebRTCInfo {
     /// WebRTC フレームワークのバージョン
-    public static let version = "M122"
+    public static let version = "M125"
 
     /// WebRTC フレームワークのコミットポジション
-    public static let commitPosition = "1"
+    public static let commitPosition = "2"
 
     /// WebRTC フレームワークのメンテナンスバージョン
-    public static let maintenanceVersion = "0"
+    public static let maintenanceVersion = "5"
 
     /// WebRTC フレームワークのソースコードのリビジョン
-    public static let revision = "6b419a0536b1a0ccfff3682f997c6f19bcbd9bd8"
+    public static let revision = "8505a9838ea91c66c96c173d30cd66f9dbcc7548"
 
     /// WebRTC の branch-heads
-    public static let branch = "6261"
+    public static let branch = "6422"
 
     /// WebRTC フレームワークのソースコードのリビジョン (短縮版)
     public static var shortRevision: String {