+++ Sora/PeerChannel.swift
@@ -195,9 +195,9 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
         lock.lock()
         onConnect = handler
 
+        // TODO(zztkm): WrapperVideoEncoderFactory は type: offer メッセージを受け取ったときに設定されるので、ここでの設定は不要かもしれない
         // サイマルキャストを利用する場合は、 RTCPeerConnection の生成前に WrapperVideoEncoderFactory を設定する必要がある
-        // また、スポットライトはサイマルキャストを利用しているため、同様に設定が必要になる
-        WrapperVideoEncoderFactory.shared.simulcastEnabled = configuration.simulcastEnabled || configuration.spotlightEnabled == .enabled
+        WrapperVideoEncoderFactory.shared.simulcastEnabled = configuration.simulcastEnabled
 
         signalingChannel.connect { [weak self] error in
             guard let weakSelf = self else {