+++ Package.swift
@@ -3,7 +3,7 @@
 import Foundation
 import PackageDescription
 
-let file = "WebRTC-125.6422.2.5/WebRTC.xcframework.zip"
+let file = "WebRTC-127.6533.1.1/WebRTC.xcframework.zip"
 
 let package = Package(
     name: "Sora",
@@ -16,7 +16,7 @@ let package = Package(
         .binaryTarget(
             name: "WebRTC",
             url: "https://github.com/shiguredo/sora-ios-sdk-specs/releases/download/\(file)",
-            checksum: "5f3b36455b72b1448800c38aae5435b973f1ad18169b9ce004e8f7eaee246be4"
+            checksum: "b9242358b4d53cafdf19d75a731cea87c84205475f54816e3a5cfd99cdb03216"
         ),
         .target(
             name: "Sora",