+++ CHANGES.md
@@ -13,6 +13,10 @@
 
 - [UPDATE] WebRTC m125.6422.2.5 に上げる
   - @miosakuma
+- [UPDATE] CocoaPods のソースリポジトリを GitHub から CDN に変更する
+  - CocoaPods 1.8 からソースリポジトリのデフォルトが `https://cdn.cocoapods.org/` になった
+  - https://blog.cocoapods.org/CocoaPods-1.8.0-beta/
+  - @zztkm
 
 ## 2024.2.0