+++ CHANGES.md
@@ -24,8 +24,9 @@
 - [FIX] `WrapperVideoEncoderFactory.shared.simulcastEnabled` の値を type: offer の際に設定される simulcast の値で上書きする
   - 認証ウェブフック成功時に払い出された type: offer の `simulcast` の値が反映されない不具合への対応
   - @zztkm
-- [FIX] `WrapperVideoEncoderFactory.shared.simulcastEnabled` の設定条件から `Configuration.spotlightEnabled` を削除する
-  - `Configuration.spotlightEnabled` はサイマルキャストを有効化するための条件ではないため削除する
+- [FIX] `Configuration.spotlightEnabled` はサイマルキャストを有効化するための条件ではないのに、判定条件に加わっていた問題を修正する
+  - `WrapperVideoEncoderFactory.shared.simulcastEnabled` の判定条件から `Configuration.spotlightEnabled` を削除する
+  - 本来は、<https://github.com/shiguredo/sora-ios-sdk/commit/44f3b81fd81694f3f670e3de568afc2a6bab5f9f> 時点で修正すべきだったが、漏れていたので修正
   - @zztkm
 
 ## 2024.2.0