+++ CHANGES.md
@@ -11,8 +11,9 @@
 
 ## develop
 
-- [UPDATE] WebRTC m125.6422.2.5 に上げる
+- [UPDATE] WebRTC m127.6533.1.1 に上げる
   - @miosakuma
+  - @zztkm
 - [UPDATE] CocoaPods のソースリポジトリを GitHub から CDN に変更する
   - CocoaPods 1.8 からソースリポジトリのデフォルトが `https://cdn.cocoapods.org/` になった
   - https://blog.cocoapods.org/CocoaPods-1.8.0-beta/