+++ CHANGES.md
@@ -11,6 +11,9 @@
 
 ## develop
 
+- [CHANGE] MediaChannelConfiguration を非推奨にする
+  - SDK 内部では利用していないため
+  - @zztkm
 - [UPDATE] WebRTC m127.6533.1.1 に上げる
   - @miosakuma
   - @zztkm